move helpers into their own file, create error config function
